+++ CHANGELOG.rst
@@ -1,3 +1,3 @@
-Changelog
--------
--------
+## Other
+- Add Changelog file+Empty script file
+